@@ -2,7 +2,9 @@ source 'https://rubygems.org'
 
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.11.1'
+gem 'rails', '5.0.1'
+# Dependence required by rails
+gem 'actionpack', '5.0.1'
 # Use SCSS for stylesheets
 gem 'sass-rails', '~> 5.0'
 # Use Uglifier as compressor for JavaScript assets
@@ -20,10 +22,10 @@ gem 'turbolinks'
 gem 'jbuilder', '~> 2.0'
 # bundle exec rake doc:rails generates the API under doc/api.
 gem 'sdoc', '~> 0.4.0', group: :doc
-gem 'devise', '~> 4.1.1'
+gem 'devise', '~> 4.7', '>= 4.7.1'
 gem 'simple_form', '~> 3.2', '>= 3.2.1'
 gem 'httparty', '~> 0.14.0' #HTTP Requests
-gem 'carrierwave', '~> 1.0.0' #Uploader
+gem 'carrierwave', '~> 2.1' #Uploader
 gem 'mini_magick', '~> 4.9', '>= 4.9.5' #Manipulates images
 gem 'carrierwave-bombshelter', '~> 0.2.2' #Protects from bomb images
 gem "recaptcha", require: "recaptcha/rails" #reCaptcha
@@ -40,7 +42,7 @@ gem "recaptcha", require: "recaptcha/rails" #reCaptcha
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug'
-  gem 'rspec-rails', '~> 3.4.2'
+  gem 'rspec-rails', '~> 3.9'
   gem 'factory_girl_rails', '~> 4.7.0'
   gem 'capybara', '~> 2.7.1'
 end
