@@ -1,3 +1,3 @@
-class EcoVariable < ActiveRecord::Base
+class EcoVariable < ApplicationRecord
 	belongs_to :eco_variables_species
 end
