@@ -1,3 +1,3 @@
-class EcoVariablesSpecies < ActiveRecord::Base
+class EcoVariablesSpecies < ApplicationRecord
 	belongs_to :species
 end
