@@ -1,4 +1,4 @@
-class GroupsUser < ActiveRecord::Base
+class GroupsUser < ApplicationRecord
 	belongs_to :user
   	belongs_to :group
 
