@@ -1,1 +1,1 @@
-BASE_URI = "http://192.168.11.42:8080/v2";
+BASE_URI = "http://192.168.11.105:4500/v2";
